requirements.txt
uvicorn==0.29.0
Pillow==10.0.1
fastapi==0.111.0
requests==2.31.0
transformers==4.40.2
python-multipart==0.0.7
torch==2.2.0